+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Research/BlueLeftPipe.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Research;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.opencv.core.Core;
@@ -9,9 +9,7 @@ import org.opencv.core.Scalar;
 import org.opencv.imgproc.Imgproc;
 import org.openftc.easyopencv.OpenCvPipeline;
 
-import java.util.TreeMap;
-
-public class ImageRecogPipe extends OpenCvPipeline {
+public class BlueLeftPipe extends OpenCvPipeline {
     Telemetry telemetry;
     int correctlocation = 3;
     Mat mat = new Mat();
@@ -21,16 +19,14 @@ public class ImageRecogPipe extends OpenCvPipeline {
         LEFT
     }
     private Location location;
-    //Defining rectangles
-    static final Rect Middle = new Rect(
-            new Point(47, 45),
-            new Point(120, 75));
-    //Defining rectangles
-    static final Rect Right = new Rect(
-            new Point(140, 35),
-            new Point(200, 75));
-    static final double PERCENT_COLOR_THRESHOLD = 0.1;
-    public ImageRecogPipe(Telemetry t) {telemetry = t;}
+    static final Rect BMiddle = new Rect(
+            new Point(200, 70),
+            new Point(250, 90));
+    static final Rect BLeft = new Rect(
+            new Point(80, 70),
+            new Point(120, 95));
+    static final double PERCENT_COLOR_THRESHOLD = 0.2;
+    public BlueLeftPipe(Telemetry t) {telemetry = t;}
 
     @Override
     public Mat processFrame(Mat input) {
@@ -40,29 +36,27 @@ public class ImageRecogPipe extends OpenCvPipeline {
 
         Core.inRange(mat,lowHSV,highHSV,mat);
 
-        //Creating the rectangles defined above
-        Mat middle = mat.submat(Middle);
-        Mat right = mat.submat(Right);
+        Mat middle = mat.submat(BMiddle);
+        Mat left = mat.submat(BLeft);
 
-
-        double middleValue = Core.sumElems(middle).val[0] / Middle.area() / 255;
-        double rightValue = Core.sumElems(right).val[0] / Right.area() / 255;
+        double middleValue = Core.sumElems(middle).val[0] / BMiddle.area() / 255;
+        double leftValue = Core.sumElems(left).val[0] / BLeft.area() / 255;
 
         middle.release();
-        right.release();
+        left.release();
 
-        telemetry.addData("Right raw value", (int) Core.sumElems(right).val[0]);
+        telemetry.addData("Right raw value", (int) Core.sumElems(left).val[0]);
         telemetry.addData("Middle raw value", (int) Core.sumElems(middle).val[0]);
-        telemetry.addData("Right percentage", Math.round(rightValue * 100) + "%");
+        telemetry.addData("Right percentage", Math.round(leftValue * 100) + "%");
         telemetry.addData("Middle percentage", Math.round(middleValue * 100) + "%");
 
 
-        boolean onRight = rightValue >PERCENT_COLOR_THRESHOLD;
+        boolean onLeft = leftValue >PERCENT_COLOR_THRESHOLD;
         boolean onMiddle = middleValue>PERCENT_COLOR_THRESHOLD;
 
-        if (onRight){
+        if (onLeft){
             correctlocation = 1;
-            telemetry.addData("LOCATION!:","RIGHT");
+            telemetry.addData("LOCATION!:","LEFT");
 
         }
         else if (onMiddle){
@@ -71,7 +65,7 @@ public class ImageRecogPipe extends OpenCvPipeline {
         }
         else{
             correctlocation = 3;
-            telemetry.addData("LOCATION!:","LEFT");
+            telemetry.addData("LOCATION!:","RIGHT");
         }
         telemetry.update();
         Scalar False = new Scalar(255, 0, 0);
@@ -79,8 +73,8 @@ public class ImageRecogPipe extends OpenCvPipeline {
 
 
         Imgproc.cvtColor(mat,mat,Imgproc.COLOR_GRAY2RGB);
-        Imgproc.rectangle(mat,Right , location == Location.RIGHT? True:False);
-        Imgproc.rectangle(mat,Middle, location == Location.MIDDLE? True :False);
+        Imgproc.rectangle(mat,BLeft , location == Location.RIGHT? True:False);
+        Imgproc.rectangle(mat,BMiddle, location == Location.MIDDLE? True :False);
         return mat;
     }
     public Location getLocation(){
