+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Research/EPIC_TeleOp_Prototype.java
@@ -2,7 +2,7 @@
 forward and backward movement and letting the right stick x control the rotating of
 the robot just like an Rc Car.
 */
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Research;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
