+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Research/allAuto.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Research;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
