+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Research/newAutonRedLeft.java
@@ -1,20 +1,15 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Research;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.util.ElapsedTime;
 
 import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.teamcode.RobotObjects.EPIC.Claw;
 import org.firstinspires.ftc.teamcode.RobotObjects.EPIC.Mecanum_Wheels;
 import org.firstinspires.ftc.teamcode.RobotObjects.Spinner;
-import org.firstinspires.ftc.teamcode.tfrec.Detector;
-import org.firstinspires.ftc.teamcode.tfrec.classification.Classifier;
 import org.openftc.easyopencv.OpenCvCamera;
 import org.openftc.easyopencv.OpenCvCameraFactory;
 import org.openftc.easyopencv.OpenCvCameraRotation;
-
-import java.util.List;
 
 @Autonomous(name="newAuton_RedLeft", group="Robot19587")
 public class newAutonRedLeft extends LinearOpMode {
